+++ db/migrate/20190611095744_create_de_identified_identifiers.rb
@@ -2,7 +2,7 @@ class CreateDeIdentifiedIdentifiers < ActiveRecord::Migration[5.2]
   def change
     create_table     :de_identified_identifiers , :primary_key => :de_identified_identifier_id do |t|
     	t.string          :identifier, null: false
-    	t.bigint          :person_id, null: false
+    	t.integer          :person_id, null: false
     	t.boolean       :voided, null: false, default: 0
     	t.bigint           :voided_by
     	t.datetime      :voided_date
